backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
PyJWT[crypto]==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
anthropic==0.26.0
supabase==2.4.6
stripe==9.5.0
twilio==9.1.0
resend==0.8.0
python-magic==0.4.27
#Pillow==10.3.0
cryptography==42.0.7
structlog==24.1.0
bleach==6.1.0
slowapi==0.1.9
tenacity==8.3.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
psycopg2-binary==2.9.9
